move listeners into SFTPDServer class
