+++ index.js
@@ -3,14 +3,10 @@ export const DEFAULT_CONF = "sftpd.conf";
 export const DEFAULT_SFTP_BANNER = "PUT it in and GET it out";
 export const DEFAULT_SFTP_PORT = 22;
 
-export {requestListener} from "./src/admin.js";
 export {default as configure} from "./src/configure.js";
-export {default as createHTTPServer} from "./src/create-http-server.js";
-export {default as createSFTPServer} from "./src/create-sftp-server.js";
 export {FTPProtocol, FTPSession, FTPRequestContext} from "./src/ftp.js";
 export * as hash from "./src/hash.js";
 export * as http from "./src/http.js";
-export {connectionListener} from "./src/sftp.js";
 export {SFTPDConsole} from "./src/sftpd-console.js";
 export {SFTPDServer} from "./src/sftpd-server.js";
 export {TemporaryStorage} from "./src/storage.js";
