--- src/create-http-server.js
@@ -1,6 +1,0 @@
-import tlsopt from "tlsopt";
-
-export default function createHTTPServer(listener) {
-  const server = tlsopt.createServerSync(listener);
-  return server;
-}
